lambda_fns/leave_reports/requirements.txt
boto3==1.36.6
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
python-dateutil==2.9.0.post0